Print the size of various types on the computer it is compiled and run on

+++ 0x00-hello_world/6-size.c
@@ -11,9 +11,9 @@
 
 int main(void)
 {
-	printf("Size of a char: %lu byte(s)\n", sizeof(char));
-	print("Size of an int: %lu byte(s)\n", sizeof(int));
-	printf("Size of a long int: %lu byte(s)\n", sizeof(long int));
-	printf("Size of a long long int: %lu byte(s)\n", sizeof(long long int));
-	printf("Size of a float: %lu byte(s)\n", sizeof(float));
+	printf("Size of a char: %lu byte(s)\n", (unsigned long)sizeof(char));
+	print("Size of an int: %lu byte(s)\n", (unsigned long)sizeof(int));
+	printf("Size of a long int: %lu byte(s)\n", (unsigned long)sizeof(long int));
+	printf("Size of a long long int: %lu byte(s)\n", (unsigned long)sizeof(long long int));
+	printf("Size of a float: %lu byte(s)\n", (unsigned long)sizeof(float));
 	return (0);
